requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
PyMuPDF
apscheduler
pytz
bcrypt==3.2.0
google-generativeai
google-cloud-aiplatform
google-cloud-texttospeech
google-cloud-speech
paypalrestsdk
pydantic[email]